requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytest==8.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
